feat(xquery): QT3 lane-15 core batch — switch/group-by/if gated (60 cases)
Vendor prod/SwitchExpr.xml, prod/GroupByClause.xml (+ GroupByUseCases
source docs and the shared docs/works-mod.xml environment) and
prod/IfExpr.xml; gate SwitchExpr 16, GroupByClause 20 (works-mod
context + docvar-bound external variables), IfExpr 24 — every case
that runs agrees.

Engine fixes the corpus drove:

- group by: bare `$k` grouping variables (key = the variable's own
  value), comma-separated multi-spec grouping partitioned on the
  key tuple, optional `as SequenceType` and `collation` specs
  (accepted; codepoint ordering), and WHERE-after-GROUP-BY
  filtering the grouped tuples (aggregates see the whole group).
- ctor-enclosed FLWOR (`<out>{ for ... group by ... }</out>`): the
  ctor is hoisted out and wraps the result sequence once — those
  clauses only run at the XQuery layer.
- attribute/namespace nodes captured into FLWOR tuples or LET
  bindings now outlive their source nodeset: xpath_nodeset_deep_copy
  clones owned synthetic members, and tuples own (and dispose)
  their captured members; the group-tuple builder clones member
  nodes likewise (QT3 group-by keys came out empty before).
- param selects evaluate on the query's context so parse-xml()
  documents stay anchored for the whole eval (external-variable
  docvars navigated as empty before).
- nested direct constructors inside enclosed expressions splice
  recursively, with ctor-aware brace matching
  (`<out>{ <g/> }</out>` failed to parse).
- fn steps (`E/string()`) directly after a filter expression map
  per member (PATH_EXPR fn-step continuation; shared
  xpath_map_fn_over helper — also fixes the `!` map freeing its
  left nodeset before mapping).
- `prefix:*` vs multiply: the NCNAME+STAR token pair is
  disambiguated from the source text (`k * 2` parsed as `k:*`).
- multi-case switch (`case A case B return R`) carried as a
  per-member-matching sequence test; xs:float registered as the
  double constructor alias; string() without arguments is
  kind-aware for attribute/text context nodes; text{(...)} joins
  the atomics with single spaces.
- xsl:number and month-shift seconds prints use %lld casts
  (long long values on 32-bit long platforms).

Excluded with rationale in the spec: duplicate grouping variable
names (XQST0089 spec divergence), declare-bound ctor navigation
and utA-vs-numeric switch matching (string-level ctor value
model), and inner-for-with-order-by inside ctor content (needs
order-by in the expression-level for — follow-up slice).

+++ src/leptris/xslt/xslt_exec.c
@@ -4457,7 +4457,8 @@ static void emit_formatted_numbers(const unsigned long long* values, int nv,
             unsigned zcp = nd_zero[i < ns ? i : ns - 1];
             int wdt = nd_width[i < ns ? i : ns - 1];
             char dec[48];
-            snprintf(dec, sizeof(dec), "%lu", values[i]);
+            snprintf(dec, sizeof(dec), "%llu",
+                     (unsigned long long)values[i]);
             size_t dl = strlen(dec);
             size_t pad = (dl < (size_t)wdt) ? (size_t)wdt - dl : 0;
             size_t co = 0;
